requirements.txt
pandas
numpy
scikit-learn
fastapi
uvicorn
streamlit
plotly
requests
python-dotenv
google-generativeai
google-auth
nltk
passlib[bcrypt]
python-jose[cryptography]
sqlalchemy
